+++ gsa/src/web/pages/reportformats/row.js
@@ -28,7 +28,6 @@ import EntityNameTableData from 'web/entities/entitynametabledata';
 import Comment from 'web/components/comment/comment';
 
 import EditIcon from 'web/entity/icon/editicon';
-import ExportIcon from 'web/components/icon/exporticon';
 import TrashIcon from 'web/entity/icon/trashicon';
 
 import IconDivider from 'web/components/layout/icondivider';
@@ -50,7 +49,6 @@ const Actions = compose(
     capabilities,
     entity,
     onReportFormatDeleteClick,
-    onReportFormatDownloadClick,
     onReportFormatEditClick,
   }) => (
     <IconDivider align={['center', 'center']} grow>
@@ -67,11 +65,6 @@ const Actions = compose(
         entity={entity}
         onClick={onReportFormatEditClick}
       />
-      <ExportIcon
-        value={entity}
-        title={_('Export Report Format')}
-        onClick={onReportFormatDownloadClick}
-      />
     </IconDivider>
   ),
 );
@@ -79,7 +72,6 @@ const Actions = compose(
 Actions.propTypes = {
   entity: PropTypes.model.isRequired,
   onReportFormatDeleteClick: PropTypes.func.isRequired,
-  onReportFormatDownloadClick: PropTypes.func.isRequired,
   onReportFormatEditClick: PropTypes.func.isRequired,
 };